git-svn-id: file:///media/LinuxData/Development64/SVN/Xfce-Theme-Manager@669 ce3ec29c-c337-42cb-bc8f-a72e3bd640e3

+++ Xfce-Theme-Manager/src/cli.cpp
@@ -140,7 +140,7 @@ int doCliThemePart(char* name,long what)
 void cliSetPaperStyle(void)
 {
 	sprintf((char*)&generalBuffer[0],"%s%i/image-style",MONITORPROP,cliMonitor);
-	setValue(XFCEDESKTOP,(char*)&generalBuffer[0],INT,(void*)cliPaperStyle);
+	setValue(XFCEDESKTOP,(char*)&generalBuffer[0],INT,(void*)(long)cliPaperStyle);
 }
